06_loadbalancers/app_v1/app/demo/requirements.txt
apache-libcloud==2.3.0
asn1crypto==0.24.0
astroid==1.6.5
awscli==1.16.122
backports.functools-lru-cache==1.5
bcrypt==3.1.4
boto==2.49.0
boto3==1.9.6
botocore==1.12.112
cachetools==3.1.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
configparser==3.5.0
cryptography==2.3.1
Django==1.11.20
docutils==0.14
enum34==1.1.6
freeze==1.0.10
futures==3.2.0
google-api-core==1.12.0
google-auth==1.6.3
google-cloud-core==1.0.2
google-cloud-error-reporting==0.31.0
google-cloud-logging==1.11.0
googleapis-common-protos==1.6.0
grpcio==1.18.0
grpcio-tools==1.18.0
gunicorn==19.9.0
idna==2.7
ipaddress==1.0.22
isort==4.3.4
Jinja2==2.10
jmespath==0.9.3
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.16.2
paramiko==2.4.1
pep8==1.7.1
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.5
pycparser==2.18
pycrypto==2.6.1
pylint==1.9.3
PyNaCl==1.2.1
python-dateutil==2.7.3
pytz==2019.1
PyYAML==3.13
requests==2.19.1
rsa==3.4.2
s3transfer==0.2.0
singledispatch==3.4.0.3
six==1.12.0
urllib3==1.23
virtualenv==16.1.0
wrapt==1.10.11
